+++ CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>19.0.109 (2020-01-10)</small>
+
+* New: checkboxGroup ([bc76cb1](https://github.com/WTW-IM/es-components/commit/bc76cb1))
+
+
+
+
+
+## <small>19.0.108 (2020-01-08)</small>
+
+* Code cleanup ([546635f](https://github.com/WTW-IM/es-components/commit/546635f))
+* Removed settings ([be7d6d8](https://github.com/WTW-IM/es-components/commit/be7d6d8))
+* Updated react-overlays to 2.1.0 ([53723e0](https://github.com/WTW-IM/es-components/commit/53723e0))
+
+
+
+
+
 ## <small>19.0.107 (2019-12-12)</small>
 
 * Update: warning sign uses circle instead of triangle ([8d60802](https://github.com/WTW-IM/es-components/commit/8d60802))